Implement authorization user

+++ app/Http/Controllers/QuestionController.php
@@ -70,7 +70,12 @@ class QuestionController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(404, 'Access Denied');
+        }
         return view('questions.edit', compact('question'));
+
+
     }
 
     /**
@@ -82,6 +87,9 @@ class QuestionController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(404, 'Access Denied');
+        }
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success',"Your Question has been updated.");
@@ -95,6 +103,9 @@ class QuestionController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(404, 'Access Denied');
+        }
         $question->delete();
 
         return redirect('/questions')->with('success', "Your question has been successfully deleted.");
